requirements.txt
python-telegram-bot==13.15
requests
bs4
schedule
python-dotenv
